refactor[FE]: Refactor location of query, mutation
