+++ frontend/src/components/organisms/Channel/ChannelDetail.tsx
