+++ frontend/src/components/organisms/ChannelLobby.tsx
@@ -1,13 +1,19 @@
+import { useNavigate } from 'react-router-dom';
+import ChannelButtons from 'components/organisms/ChannelButtons';
 import ChannelList from 'components/molecule/ChannelList';
 import { ChannelType } from 'types';
-import ChannelButtons from 'components/organisms/ChannelButtons';
 
 interface Props {
-  channels: ChannelType[] | null;
-  onClick: (channel: ChannelType) => void;
+  channels: ChannelType[];
 }
 
-export default function ChannelLobby({ channels, onClick }: Props) {
+export default function ChannelLobby({ channels }: Props) {
+  const navigate = useNavigate();
+
+  const handleClickChannel = (channel: ChannelType) => {
+    navigate(`/channel/${channel.id}`);
+  };
+
   return (
     <>
       <section className="section">
@@ -16,8 +22,8 @@ export default function ChannelLobby({ channels, onClick }: Props) {
       </section>
       <section className="section">
         <h2 className="section-title">입장 중인 채널 목록</h2>
-        {channels?.length ? (
-          <ChannelList channels={channels} onClick={onClick} />
+        {channels.length ? (
+          <ChannelList channels={channels} onClick={handleClickChannel} />
         ) : (
           <div>입장 중인 채널이 없습니다.</div>
         )}
