+++ frontend/src/components/organisms/Profile/index.tsx
@@ -1,5 +1,5 @@
 import { Link, useNavigate } from 'react-router-dom';
-import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
+import { useMutation, useQueryClient } from '@tanstack/react-query';
 import * as api from 'api/api.v1';
 import { useUser } from 'hooks/useStore';
 import Button from 'components/atoms/Button';
@@ -9,21 +9,16 @@ import GameHistory from 'components/organisms/Profile/GameHistory';
 import { UserType } from 'types';
 
 interface Props {
-  userId: number;
+  user: UserType;
 }
 
-export default function Profile({ userId }: Props) {
+export default function Profile({ user }: Props) {
   const myUserId = useUser((state) => state.id);
   const queryClient = useQueryClient();
   const navigate = useNavigate();
 
-  const queryKey = ['profile', userId];
-  const isMyPage = userId === myUserId;
-
-  const { data: user } = useQuery({
-    queryKey,
-    queryFn: () => api.getUser(Number(userId)),
-  });
+  const queryKey = ['profile', user.id];
+  const isMyPage = user.id === myUserId;
 
   const onMutate = async (newUser: UserType | undefined) => {
     await queryClient.cancelQueries(queryKey);
@@ -74,79 +69,74 @@ export default function Profile({ userId }: Props) {
   return (
     <>
       <section className="section">
-        {user && (
-          <div className="flex flex-col items-center">
-            <ProfileImage userId={userId} alt="profile image" size={192} />
-            <div className="mb-8 text-center">
-              <p
-                data-testid={userId}
-                className="text-2xl font-semibold mt-4 mb-2"
-              >
-                {user.nickname}
-              </p>
-              <span className="badge">Lv. {user.level}</span>
-            </div>
-            {isMyPage && (
-              <Link to="/setting/profile" className="button primary small">
-                프로필 수정하기
-              </Link>
-            )}
-            {!isMyPage && (
-              <div className="flex flex-col items-center">
-                <div className="inline-flex gap-4 mb-4">
-                  {user.isFollowedByMyself ? (
-                    <Button
-                      onClick={() => unfollowUserMutation.mutate(userId)}
-                      secondary
-                      className="w-36"
-                    >
-                      언팔로우
-                    </Button>
-                  ) : (
-                    <Button
-                      onClick={() => followUserMutation.mutate(userId)}
-                      primary
-                      className="w-36"
-                    >
-                      팔로우하기
-                    </Button>
-                  )}
-                  {!user.isBlockedByMyself && (
-                    <Button
-                      onClick={() => getDmChannelMutation.mutate(userId)}
-                      primary
-                      className="w-36"
-                    >
-                      메시지 보내기
-                    </Button>
-                  )}
-                </div>
-                {user.isBlockedByMyself ? (
+        <div className="flex flex-col items-center">
+          <ProfileImage userId={user.id} alt="profile image" size={192} />
+          <div className="mb-8 text-center">
+            <p data-testid={user} className="text-2xl font-semibold mt-4 mb-2">
+              {user.nickname}
+            </p>
+            <span className="badge">Lv. {user.level}</span>
+          </div>
+          {isMyPage && (
+            <Link to="/setting/profile" className="button primary small">
+              프로필 수정하기
+            </Link>
+          )}
+          {!isMyPage && (
+            <div className="flex flex-col items-center">
+              <div className="inline-flex gap-4 mb-4">
+                {user.isFollowedByMyself ? (
                   <Button
-                    onClick={() => unblockUserMutation.mutate(userId)}
-                    linkStyle
-                    className="text-red"
-                    size="small"
+                    onClick={() => unfollowUserMutation.mutate(user.id)}
+                    secondary
+                    className="w-36"
                   >
-                    차단 해제
+                    언팔로우
                   </Button>
                 ) : (
                   <Button
-                    onClick={() => blockUserMutation.mutate(userId)}
-                    linkStyle
-                    className="text-red"
-                    size="small"
+                    onClick={() => followUserMutation.mutate(user.id)}
+                    primary
+                    className="w-36"
                   >
-                    차단하기
+                    팔로우하기
+                  </Button>
+                )}
+                {!user.isBlockedByMyself && (
+                  <Button
+                    onClick={() => getDmChannelMutation.mutate(user.id)}
+                    primary
+                    className="w-36"
+                  >
+                    메시지 보내기
                   </Button>
                 )}
               </div>
-            )}
-          </div>
-        )}
+              {user.isBlockedByMyself ? (
+                <Button
+                  onClick={() => unblockUserMutation.mutate(user.id)}
+                  linkStyle
+                  className="text-red"
+                  size="small"
+                >
+                  차단 해제
+                </Button>
+              ) : (
+                <Button
+                  onClick={() => blockUserMutation.mutate(user.id)}
+                  linkStyle
+                  className="text-red"
+                  size="small"
+                >
+                  차단하기
+                </Button>
+              )}
+            </div>
+          )}
+        </div>
       </section>
-      <Achievements userId={Number(userId)} />
-      <GameHistory userId={Number(userId)} />
+      <Achievements userId={user.id} />
+      <GameHistory userId={user.id} />
     </>
   );
 }
